+++ hashtagmap-admin/build.gradle
@@ -2,7 +2,6 @@ dependencies {
     implementation project(':hashtagmap-core')
     implementation project(':hashtagmap-kakao-scheduler')
     implementation project(':hashtagmap-instagram-scheduler')
-    implementation project(':hashtagmap-crawling-scheduler')
 
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
