--- hashtagmap-core/src/main/java/com/songpapeople/hashtagmap/place/repository/PlaceRepository.java
@@ -1,8 +1,0 @@
-package com.songpapeople.hashtagmap.place.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.songpapeople.hashtagmap.place.domain.model.Place;
-
-public interface PlaceRepository extends JpaRepository<Place, Long> {
-}
