+++ hashtagmap-core/src/main/resources/application.yml
@@ -10,7 +10,7 @@ spring:
       ddl-auto: create
     properties:
       hibernate:
-        dialect: org.hibernate.dialect.MySQL5InnoDBDialect
+        dialect: org.hibernate.dialect.H2Dialect
         format_sql: true
     show-sql: true
     generate-ddl: true
