--- hashtagmap-core/src/test/java/com/songpapeople/hashtagmap/CoreApplicationTest.java
@@ -1,12 +1,0 @@
-package com.songpapeople.hashtagmap;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CoreApplicationTest {
-
-    @Test
-    void contextLoads() {
-    }
-}
