+++ settings.gradle
@@ -3,10 +3,8 @@ include 'hashtagmap-web'
 include 'hashtagmap-core'
 include 'hashtagmap-instagram-crawler'
 include 'hashtagmap-kakao-api'
-include 'hashtagmap-instagram-api'
 include 'hashtagmap-admin'
 include 'hashtagmap-kakao-scheduler'
 include 'hashtagmap-instagram-scheduler'
-include 'hashtagmap-crawling-scheduler'
 include 'hashtagmap-common'
